requirements.txt
Django==5.2.9
paho-mqtt
celery
redis
django-celery-beat
djangorestframework
channels
channels-redis
daphne
websockets
